Backend get task by ID

+++ backend/case-management/src/main/java/uk/gov/hmcts/case_management/repository/TaskRepository.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.case_management.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import uk.gov.hmcts.case_management.model.Task;
 
-public interface TaskRepository extends JpaRepository<Task, Integer>{}
+public interface TaskRepository extends JpaRepository<Task, Long>{}
